some prelim cleanups
